Remove ClientIntegrationTests superclass again
